EasyJourney Project
This project is based on a live train tracking system for the Indian Railways.
I could only make the frontend for this website as I was having difficulty in understanding and implementing the backend. I hope that I'll be able to learn further and make this a full stack website.